+++ backend/src/main/java/com/rssmanager/auth/application/AuthService.java
@@ -1,4 +1,4 @@
-package com.rssmanager.auth.service;
+package com.rssmanager.auth.application;
 
 import com.rssmanager.auth.controller.dto.CertificateResponse;
 import com.rssmanager.auth.controller.dto.LoginRequest;
