+++ backend/src/test/java/com/rssmanager/support/MockGithubExchange.java
@@ -1,6 +1,6 @@
 package com.rssmanager.support;
 
-import com.rssmanager.auth.service.OAuthExchangeService;
+import com.rssmanager.auth.application.OAuthExchangeService;
 import com.rssmanager.member.domain.Member;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
